marianatek/requirements.txt
django
django-extensions
django-filter
djangorestframework
ipython
requests
Werkzeug
